requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
Faker==4.0.3
idna==2.9
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
pytest-rerunfailures==9.0
python-dateutil==2.8.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.1.9
webdriver-manager==3.2.0
